pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trading-platform-binance-pg-v8"
version = "0.8.0"
requires-python = ">=3.11"
dependencies = [
  "pyyaml>=6.0.1",
  "requests>=2.31.0",
  "websocket-client>=1.8.0",
  "psycopg[binary]>=3.2.1",
  "psycopg-pool>=3.2.1",
  "python-dotenv>=1.0.1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
